refactor(swipe-card): replace useMounted with next/dynamic ssr:false for AnimatedCard
