+++ frontend/src/components/learn/swipe-card.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import { animated, useSpring } from "@react-spring/web";
-import { useDrag } from "@use-gesture/react";
+import dynamic from "next/dynamic";
 import { RotateCcw, Smile, Zap } from "lucide-react";
-import { useCallback, useEffect, useState } from "react";
 import type { SwipeDirection } from "@/app/learn/[cardgroupId]/learn-client";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
@@ -33,115 +31,15 @@ const directionMeta: Record<
   right: { label: "Easy", icon: Smile, className: "text-emerald-700 hover:bg-emerald-50" },
 };
 
-function useMounted() {
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => setMounted(true), []);
-  return mounted;
-}
+// Load AnimatedCard only on the client to avoid @react-spring/web hydration mismatches.
+// The loading fallback renders nothing (same as the old useMounted() === false branch).
+const AnimatedCard = dynamic(
+  () => import("./animated-card").then((m) => m.AnimatedCard),
+  { ssr: false },
+);
 
-function StaticCard({ card, isActive, onSwipe }: Omit<Props, "onSwipeProgress">) {
-  return (
-    <article
-      className="absolute inset-0 touch-none select-none"
-      aria-label={`Flashcard: ${card.front}`}
-      tabIndex={isActive ? 0 : -1}
-    >
-      <CardContent card={card} isActive={isActive} onSwipe={onSwipe} />
-    </article>
-  );
-}
-
-function AnimatedCard({ card, isActive, onSwipe, onSwipeProgress }: Props) {
-  const [{ x, y, rotate, scale }, api] = useSpring(() => ({
-    x: 0,
-    y: 0,
-    rotate: 0,
-    scale: 1,
-    config: { tension: 520, friction: 38 },
-  }));
-
-  const completeSwipe = useCallback(
-    (direction: SwipeDirection) => {
-      const flyX =
-        direction === "left" ? -window.innerWidth : direction === "right" ? window.innerWidth : 0;
-      const flyY = direction === "down" ? window.innerHeight : 0;
-      api.start({
-        x: flyX,
-        y: flyY,
-        rotate: direction === "left" ? -16 : direction === "right" ? 16 : 0,
-        scale: 0.92,
-      });
-      onSwipe(card, direction);
-    },
-    [api, card, onSwipe],
-  );
-
-  const bind = useDrag(
-    ({ active, movement: [mx, my], direction: [, yDir], velocity: [vx, vy] }) => {
-      if (!isActive) return;
-
-      const xAbs = Math.abs(mx);
-      const yAbs = Math.abs(my);
-      let direction: SwipeDirection | null = null;
-      let progress = 0;
-
-      if (xAbs >= yAbs && xAbs > 8) {
-        direction = mx < 0 ? "left" : "right";
-        progress = Math.min(xAbs / 120, 1);
-      } else if (my > 8) {
-        direction = "down";
-        progress = Math.min(yAbs / 96, 1);
-      }
-      onSwipeProgress?.(active ? direction : null, active ? progress : 0);
-
-      const shouldSwipe =
-        !active &&
-        direction &&
-        ((direction === "down" && (yAbs > 96 || vy > 0.35 || yDir > 0.9)) ||
-          (direction !== "down" && (xAbs > 120 || vx > 0.45)));
-
-      if (shouldSwipe && direction) {
-        completeSwipe(direction);
-        return;
-      }
-
-      api.start({
-        x: active ? mx : 0,
-        y: active ? Math.max(my, -40) : 0,
-        rotate: active ? mx / 16 : 0,
-        scale: active ? 1.02 : 1,
-        immediate: active,
-      });
-    },
-    {
-      axis: undefined,
-      filterTaps: true,
-      pointer: { capture: false },
-      rubberband: true,
-      threshold: 4,
-    },
-  );
-
-  return (
-    <animated.article
-      {...bind()}
-      className="absolute inset-0 touch-none select-none cursor-grab active:cursor-grabbing"
-      aria-label={`Flashcard: ${card.front}`}
-      data-testid="swipe-card"
-      tabIndex={isActive ? 0 : -1}
-      style={{
-        x,
-        y,
-        rotate: rotate.to((value) => `${value}deg`),
-        scale,
-      }}
-    >
-      <CardContent card={card} isActive={isActive} onSwipe={onSwipe} />
-    </animated.article>
-  );
-}
-
-function CardContent({ card, isActive, onSwipe }: Omit<Props, "onSwipeProgress">) {
+// CardContent is exported so animated-card.tsx can share the same presentational layer.
+export function CardContent({ card, isActive, onSwipe }: Omit<Props, "onSwipeProgress">) {
   return (
     <div className="flex h-full flex-col rounded-lg border border-border bg-card p-6 shadow-lg">
       <div className="flex flex-1 flex-col items-center justify-center gap-8 text-center">
@@ -178,8 +76,5 @@ function CardContent({ card, isActive, onSwipe }: Omit<Props, "onSwipeProgress">
 }
 
 export function SwipeCard(props: Props) {
-  const mounted = useMounted();
-  if (!mounted)
-    return <StaticCard card={props.card} isActive={props.isActive} onSwipe={props.onSwipe} />;
   return <AnimatedCard {...props} />;
 }
